backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==3.2.2
jinja2
slowapi
razorpay
alembic